refactor(web): extract shared image upload helper
Move the photo validation, resize and hex-colour helpers out of the form
builder's ImageUpload into lib/imageUpload, so the document editor's photo
dialog can reuse them rather than duplicate them. components/builder is
removed by a later issue, which is why the helper cannot live there.
