+++ apps/web/src/components/builder/ImageUpload.tsx
@@ -1,120 +1,18 @@
 import { Show, createSignal } from "solid-js";
 import { Button, Switch, toast } from "../ui";
+import {
+  ACCEPTED_IMAGE_TYPES,
+  expandShortHex,
+  isHexColor,
+  normalizeHexColor,
+  processImage,
+  validateImageFile,
+} from "../../lib/imageUpload";
 import type { Picture } from "../../wasm/types";
 
 export interface ImageUploadProps {
   picture: Picture;
   onPictureChange: (picture: Picture) => void;
-}
-
-/** Resume photos display at ≤200px; keep the encoded asset compact. */
-const MAX_SIZE_PX = 512;
-const ACCEPTED_TYPES = ["image/jpeg", "image/png", "image/webp"];
-const MAX_FILE_SIZE = 5 * 1024 * 1024; // 5 MB
-/** Soft cap for the data URL stored in resume JSON (~300 KB payload). */
-const MAX_DATA_URL_CHARS = 400_000;
-
-const HEX_COLOR_REGEX = /^#(?:[0-9A-Fa-f]{3,4}|[0-9A-Fa-f]{6}|[0-9A-Fa-f]{8})$/;
-const BARE_HEX_REGEX = /^(?:[0-9A-Fa-f]{3,4}|[0-9A-Fa-f]{6}|[0-9A-Fa-f]{8})$/;
-
-/** Trim and prepend `#` when the value is bare 3/4/6/8-digit hex. */
-function normalizeHexColor(raw: string): string {
-  const trimmed = raw.trim();
-  return BARE_HEX_REGEX.test(trimmed) ? `#${trimmed}` : trimmed;
-}
-
-/** Expand #RGB / #RGBA shorthand to #RRGGBB / #RRGGBBAA (server validation only accepts long form). */
-function expandShortHex(color: string): string {
-  const digits = color.slice(1);
-  if (digits.length === 3 || digits.length === 4) {
-    return `#${digits
-      .split("")
-      .map((c) => c + c)
-      .join("")}`;
-  }
-  return color;
-}
-
-interface ProcessedImage {
-  dataUrl: string;
-  aspectRatio: number;
-}
-
-/**
- * Encode canvas to a data URL, preferring WebP and stepping quality down until
- * the result fits under {@link MAX_DATA_URL_CHARS}.
- */
-function encodeCanvasDataUrl(canvas: HTMLCanvasElement): string {
-  const qualities = [0.82, 0.7, 0.58, 0.45];
-  let best = "";
-  for (const quality of qualities) {
-    let dataUrl = canvas.toDataURL("image/webp", quality);
-    if (!dataUrl.startsWith("data:image/webp")) {
-      dataUrl = canvas.toDataURL("image/jpeg", quality);
-    }
-    best = dataUrl;
-    if (dataUrl.length <= MAX_DATA_URL_CHARS) {
-      return dataUrl;
-    }
-  }
-  return best;
-}
-
-/**
- * Resize an image file to fit within MAX_SIZE_PX x MAX_SIZE_PX,
- * then return a base64 data URL (WebP with JPEG fallback) and aspect ratio.
- */
-async function processImage(file: File): Promise<ProcessedImage> {
-  return new Promise((resolve, reject) => {
-    const url = URL.createObjectURL(file);
-    const img = new Image();
-    img.onload = () => {
-      URL.revokeObjectURL(url);
-      const canvas = document.createElement("canvas");
-
-      let { width, height } = img;
-      if (width > MAX_SIZE_PX || height > MAX_SIZE_PX) {
-        if (width > height) {
-          height = Math.max(1, Math.round((height * MAX_SIZE_PX) / width));
-          width = MAX_SIZE_PX;
-        } else {
-          width = Math.max(1, Math.round((width * MAX_SIZE_PX) / height));
-          height = MAX_SIZE_PX;
-        }
-      }
-
-      canvas.width = width;
-      canvas.height = height;
-      const ctx = canvas.getContext("2d");
-      if (!ctx) {
-        reject(new Error("Could not get canvas context"));
-        return;
-      }
-      ctx.drawImage(img, 0, 0, width, height);
-
-      const dataUrl = encodeCanvasDataUrl(canvas);
-      if (dataUrl.length > MAX_DATA_URL_CHARS) {
-        reject(new Error("Image is still too large after compression. Try a smaller photo."));
-        return;
-      }
-      resolve({ dataUrl, aspectRatio: width / height });
-    };
-    img.onerror = () => {
-      URL.revokeObjectURL(url);
-      reject(new Error("Failed to load image"));
-    };
-    img.src = url;
-  });
-}
-
-function validateFile(file: File): string | null {
-  if (!ACCEPTED_TYPES.includes(file.type)) {
-    return "Please upload a JPG, PNG, or WebP image.";
-  }
-  if (file.size > MAX_FILE_SIZE) {
-    return "Image must be smaller than 5 MB.";
-  }
-  return null;
 }
 
 export function ImageUpload(props: ImageUploadProps) {
@@ -126,7 +24,7 @@ export function ImageUpload(props: ImageUploadProps) {
   const hasPhoto = () => props.picture.url !== "";
 
   async function handleFile(file: File) {
-    const error = validateFile(file);
+    const error = validateImageFile(file);
     if (error) {
       toast.error(error);
       return;
@@ -209,7 +107,7 @@ export function ImageUpload(props: ImageUploadProps) {
 
   function updateColorEffect(key: "borderColor" | "shadowColor", raw: string) {
     const normalized = normalizeHexColor(raw);
-    if (normalized === "" || HEX_COLOR_REGEX.test(normalized)) {
+    if (normalized === "" || isHexColor(normalized)) {
       updateEffect(key, expandShortHex(normalized));
     }
   }
@@ -266,7 +164,7 @@ export function ImageUpload(props: ImageUploadProps) {
       <input
         ref={(el) => (fileInputRef = el)}
         type="file"
-        accept={ACCEPTED_TYPES.join(",")}
+        accept={ACCEPTED_IMAGE_TYPES.join(",")}
         class="hidden"
         aria-label="Choose profile photo"
         onChange={handleInputChange}
